+++ .gitignore
@@ -37,4 +37,4 @@ yarn-error.log*
 *.tsbuildinfo
 
 # uri
-.config.ts
+/api/core/index.ts
