+++ components/Loading/Loading.tsx
@@ -1,0 +1,5 @@
+import { Text } from './LoadingStyle';
+
+export const Loading = () => {
+  return <Text>Loading...</Text>;
+};
