+++ components/Loading/LoadingStyle.tsx
@@ -1,0 +1,6 @@
+import styled from 'styled-components';
+
+export const Text = styled.div`
+  text-align: center;
+  font-size: 5rem;
+`;
