--- config.ts
@@ -1,1 +1,0 @@
-export const URI = 'http://dev.pixel.sc:1338';
