+++ pages/_app.tsx
@@ -6,9 +6,11 @@ import { useRouter } from 'next/router';
 import GlobalStyle from '../styles/GlobalStyle';
 import Theme from '../styles/Theme';
 import Header from '../components/Header/Header';
+import axios from 'axios';
 
 function MyApp({ Component, pageProps }: AppProps) {
   const router = useRouter();
+  axios.defaults.baseURL = 'http://dev.pixel.sc:1338';
 
   return (
     <>
